+++ src/gemseo_benchmark/scenario.py
@@ -175,17 +175,19 @@ class Scenario:
             log_gemseo_to_file: Whether to save the GEMSEO log to a file
                 next to the performance history file.
         """
+        algorithms_configurations = AlgorithmsConfigurations()
+        for group in self._algorithms_configurations_groups:
+            for algorithm_configuration in group:
+                if algorithm_configuration not in algorithms_configurations:
+                    algorithms_configurations.add(algorithm_configuration)
+
         Benchmarker(
             self._histories_path,
             self._results_path,
             self._get_dir_path(self.__DATABASES_DIRNAME) if save_databases else None,
         ).execute(
             {problem for group in problems_groups for problem in group},
-            AlgorithmsConfigurations(*[
-                algo_config
-                for algos_configs_group in self._algorithms_configurations_groups
-                for algo_config in algos_configs_group
-            ]),
+            algorithms_configurations,
             overwrite_histories,
             number_of_processes,
             use_threading,